requirements.txt
Django==3.1.6
djangorestframework-jwt==1.11.0
djangorestframework==3.12.2
djangorestframework-simplejwt
django-allauth==0.44.0
Faker==6.1.1